pyproject.toml
[project]
name = "hello-chainlit"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "saadahmedacademy", email = "saadqurashiazeemqurashi@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "chainlit>=2.5.5",
    "openai-agents>=0.0.16",
]

[project.scripts]
hello-chainlit = "hello_chainlit:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
